+++ scripts/find_and_replace.gd
@@ -1,0 +1,4 @@
+extends ActionScript
+
+func _run_action() -> void:
+	Signals.open_find_panel.emit()