+++ scripts/find_previous.gd
@@ -1,0 +1,4 @@
+extends ActionScript
+
+func _run_action() -> void:
+	Signals.shift_find_result.emit(false)