+++ scripts/replace_all.gd
@@ -1,0 +1,4 @@
+extends ActionScript
+
+func _run_action() -> void:
+	Signals.replace_all.emit()