+++ CHANGELOG.md
@@ -16,6 +16,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Translation system
 - Settings sytstem
 - Log editor notifications
+- Find & Replace panel
+- Action Scripts for:
+	- Edit:
+		- Evulate Selection
+	- Search:
+		- Find and Replace
+		- Find Next
+		- Find Previous
+		- Replace All
+	- Command:
+		- Remove All Indents
+	- View:
+		- Change Editor Font Size
+		- Focus On Editor
+		- Breakpoints
+		- Bookmarks
+		- Line Numbers
+	- Help:
+		- Github Repository
+		- Mode Library
+		- Forum
+		- Share Feedback
+		- Submit Issue
+		- Online Docs
+		- About
 
 ### Removed
 
@@ -23,7 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- [Fix call _update_recent_files from scripts](https://github.com/mkh-user/Text-Forge/commit/8c8ce45f759af6887699f7702c8abfde04915a98).
+- [Fix call _update_recent_files from scripts](https://github.com/mkh-user/Text-Forge/commit/8c8ce45f759af6887699f7702c8abfde04915a98)
 
 ## [0.1.0] - 2025-06-30